requirements.txt
alpaca-py
matplotlib
pandas
pandas-datareader # not sure what this is for 
quantstats # not sure what this is for
requests # probably remove later
yfinance # remove later
lumibot 
numpy
ruptures
imageio # remove later 
json 
yaml

#NOTE(pin versions later) 